修正しました。

+++ PG2_評価課題2/Enemy.h
@@ -10,7 +10,7 @@
 
 const int ENEMY_NUM = 1;//エネミーの数
 
-const float ENEMY_SIZE = 32;//エネミーのピクセル数
+const int ENEMY_SIZE = 32;//エネミーのピクセル数
 //エネミーに使う型
 typedef struct EnemyObject {
 	Vertex local;
